django_telegrambot/requirements.txt
django==3.2.2
python-telegram-bot==13.5
emoji==1.2.0